Add basic move and reset functionality to scoreboard
Needed to fix declaring the variables for cardOrder and cardList to within
the initializeDeck function. This function also needed to reset the card classNames.
Starting the game was wrapped into a startOrResetGame function which calls moves,
stars, deck, and eventually timer. Multiple calls at the bottom of the script was
put into this single wrapper function to be called at run-time and whenever
the reset button is clicked.